Remove distribute hard-coded version
Distribute is not used within the project, so it should not be a  part of the setup.py requirements. 

There should not be a situation where it is required to be hard-coded like that. If there is due to some distribute bug referencing other deployment options used by your company - it should be set as minimal version rather than set version given that you would use setup.py based on your company setup rather than leaving project clean.