Fail Jenkins check if job is unknown
Currently, if there is there is no Jenkins job matching
the Cabot check, the check always succeeds.

This is particularly bad when jobs are renamed on Jenkins,
as Cabot stops monitoring it and never alerts on this.

Although Cabot is currently very relaxed when dealing with
Jenkins errors, since 94dcdf9 Cabot did fail on missing job.
This must have stopped working following a Jenkins API change.

Here, we catch the `jenkinsapi` `UnknownJob` exception and
set `status_code` to the value expected by the `run` method.
Also add a unit-test for that behaviour.

This is a quick fix ; but in general the entire run method
may need refactoring to not use `status_code` anymore,
We may want to consider dealiing with the `jenkinsapi` exceptions
in the `run` method rather than catching them in `get_job_status`
and setting the `status_code`.

Closes #537