Allow multiple Jenkins backends for JenkinsStatusCheck
We added a JenkinsConfig model to store configs (atm only editable in
django admin - to be moved to some sort of plugins options page)

If no JenkinsConfig are available, one will be created from the current
environment variables (JENKINS_URL etc)

Refactor JenkinsStatusCheck to have its own table instead of being a
proxy model. This is because we're moving to a plugin architecture
for status checks so we can't keep adding every new field to the base
model.