Fix JenkinsCheck result status for in-progress builds
If a build is in progress, then every JenkinsCheck was marked as failed.

Here we fetch the last_completed build, instead of the last_build,
because that is really what we are interested in, and we are not
relying on last_build to figure out whether a build is in progress.

Closes #551