Add run-book notes to cabot.
Summary: Add runbook field so that cabot checks have a run book.  This seems to work, but there may be some fit and finish to make it look nicer. Mainly it would be good to expand the size of the input box.

Test Plan: Tested locally. UT's pass, save for MYSQL, which also fails on master.

Reviewers: elainearbaugh

Subscribers: #consumer

JIRA Issues: INFRA-4358

Differential Revision: https://phabricator.team.affirm.com/D6640