Don't uset created_by when user is AnonymousUser (dev only)
In dev mode, DISABLE_LOGIN is set to True and a not-logged-in dev will have their user set to an instance of AnonymousUser.
AnonymousUser can't be assigned to model.created_by because it's not a "real" user, causing a ValueError when creating a check (again, only if not logged in).