Pin flake8 dependencies to last known working versions
The Travis-CI build has been failing for ages because of flake8:
as the flake8 dependencies were not pinned, the codebase is failing
when using more recent versions of flake8 (or of the underlying pycodestyle).

This pins the dependencies in the tox file to
the last known working versions, as found in [1].

(This is a stop-gap measure, ideally the new violations would be fixed).

[1] https://travis-ci.org/github/arachnys/cabot/builds/424842858